requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
bitarray==3.3.1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.2.0
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
Markdown==3.8
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pillow==11.2.1
pluggy==1.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybloom_live==4.0.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytest==8.3.5
pytest-django==4.11.1
python-decouple==3.8
python-dotenv==1.1.0
redis==6.0.0
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==80.0.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
xxhash==3.5.0
zope.interface==7.2
